Factor out Python-specific changes from apidoc_formatting.py
